+++ RELEASES.md
@@ -26,6 +26,7 @@
 - Fix reg_div function compatibility with numpy in `ot.unbalanced.lbfgsb_unbalanced` via new function `ot.utils.fun_to_numpy` (PR #731)
 - Added to each example in the examples gallery the information about the release version in which it was introduced (PR #743)
 - Removed release information from quickstart guide (PR #744)
+- Implement batch parallel solvers in ot.batch (PR #745)
 - Update REAMDE with new API and reorganize examples (PR #754)
 - Speedup and update tests and wheels (PR #759)